requirements.txt
fastapi
uvicorn[standard]
asyncpg
databases
pillow
sentence-transformers
numpy
requests
psycopg2-binary
python-multipart